pyproject.toml
[project]
name = "pptx-auto-report"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.3.3",
    "python-pptx>=1.0.2",
]
